rename source file to better match its contents
