+++ src/main.js
@@ -3,7 +3,7 @@ import Vue from 'vue'
 import App from './App.vue'
 import store from './store'
 import router from './router'
-import ControlPointEventBus from './upnp_event_bus.js'
+import ControlPointEventBus from './controlpoint-event-bus'
 
 Vue.config.productionTip = false
